+++ Release_Notes.md
@@ -1,3 +1,9 @@
+# 1.5.2 - Deep Sleep Hook
+* Rehide System HUD after sleep / monitor changes / lid changes
+* Fixed side grow in/out animation that would sometimes get stuck at half
+* Fixed an issue that prevented internal monitor brightness retrieval after an external monitor was connected
+* Fixed an issue that was preventing the huds from being shown when changing to a full screen application
+
 # 1.5.1 - Background bar
 * Better visibility of the bar by adding custom shadows